requirements.txt
soundfile==0.12.1
torch==1.13.1
torchaudio==0.13.1
pytorch-lightning==1.9.2
musdb==0.4.0
numpy==1.23.5
hydra-core==1.3.1
omegaconf==2.3.0
tensorboard==2.12.0
tqdm==4.64.1
museval==0.4.0